+++ database/migrations/2025_12_02_074608_create_detail_komponen_table.php
@@ -10,19 +10,21 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('detail_komponen', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('knmp_id')->constrained('profile_knmp')->onDelete('cascade');
+        if (!Schema::hasTable('detail_komponen')) {
+            Schema::create('detail_komponen', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('knmp_id')->constrained('profile_knmp')->onDelete('cascade');
 
-            $table->string('jenis_komponen', 255);
-            $table->integer('target_unit')->nullable();
-            $table->decimal('progress_persen', 5, 2)->nullable(); // Menggunakan Progress (%) dari kuesioner
-            $table->bigInteger('anggaran')->nullable(); // Anggaran (Rp)
-            $table->bigInteger('realisasi_anggaran')->nullable();
-            $table->decimal('persen_realisasi_anggaran', 5, 2)->nullable();
-            $table->text('keterangan')->nullable();
-            $table->timestamps();
-        });
+                $table->string('jenis_komponen', 255);
+                $table->integer('target_unit')->nullable();
+                $table->decimal('progress_persen', 5, 2)->nullable(); // Menggunakan Progress (%) dari kuesioner
+                $table->bigInteger('anggaran')->nullable(); // Anggaran (Rp)
+                $table->bigInteger('realisasi_anggaran')->nullable();
+                $table->decimal('persen_realisasi_anggaran', 5, 2)->nullable();
+                $table->text('keterangan')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
