+++ database/migrations/2025_12_03_033525_create_informasi_pemasaran_table.php
@@ -8,30 +8,32 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('informasi_pemasaran', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('informasi_pemasaran')) {
+            Schema::create('informasi_pemasaran', function (Blueprint $table) {
+                $table->id();
 
-            $table->unsignedBigInteger('knmp_id');
-            $table->unsignedBigInteger('responden_id');
+                $table->unsignedBigInteger('knmp_id');
+                $table->unsignedBigInteger('responden_id');
 
-            $table->text('kendala_pemasaran_text')->nullable();
-            $table->text('cara_penanganan_ikan')->nullable();
+                $table->text('kendala_pemasaran_text')->nullable();
+                $table->text('cara_penanganan_ikan')->nullable();
 
-            $table->timestamps();
+                $table->timestamps();
 
-            // =========================
-            // FOREIGN KEY
-            // =========================
-            $table->foreign('knmp_id')
-                ->references('id')
-                ->on('knmp')
-                ->onDelete('cascade');
+                // =========================
+                // FOREIGN KEY
+                // =========================
+                $table->foreign('knmp_id')
+                    ->references('id')
+                    ->on('knmp')
+                    ->onDelete('cascade');
 
-            $table->foreign('responden_id')
-                ->references('id')
-                ->on('informasi_responden')
-                ->onDelete('cascade');
-        });
+                $table->foreign('responden_id')
+                    ->references('id')
+                    ->on('informasi_responden')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
